backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
pyserial==3.5
websockets>=13.0
google-genai>=0.2.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
